Unify highlight colors across editor and modal
